优化存储结构

+++ Source/cpp/MLNKVUtility.h
@@ -68,7 +68,7 @@ static inline double MLNKVUInt64ToDouble(uint64_t v) {
 }
 
 
-static inline int16_t MLNKVRawVarUInt32Size(uint32_t value) {
+static inline uint8_t MLNKVRawVarUInt32Size(uint32_t value) {
     if ((value & (0xffffffff << 8)) == 0) {
         return 1;
     } else if ((value & (0xffffffff << 16)) == 0) {
@@ -79,7 +79,7 @@ static inline int16_t MLNKVRawVarUInt32Size(uint32_t value) {
     return 4;
 }
 
-static inline int16_t MLNKVRawVarUInt64Size(uint64_t value) {
+static inline uint8_t MLNKVRawVarUInt64Size(uint64_t value) {
     if ((value & (0xffffffffffffffffL << 8)) == 0) {
         return 1;
     } else if ((value & (0xffffffffffffffffL << 16)) == 0) {
@@ -102,9 +102,9 @@ static inline int16_t MLNKVRawVarUInt64Size(uint64_t value) {
 //    return 1;
 //}
 
-static const int16_t MLNKV32Size = 4;
+static const uint8_t MLNKV32Size = 4;
 
-static const int16_t MLNKV64Size = 8;
+static const uint8_t MLNKV64Size = 8;
 //
 ////// 判断机器模式 大小端
 /////**
@@ -120,32 +120,32 @@ static const int16_t MLNKV64Size = 8;
 //    return (c.b == 1);
 //}
 
-static inline uint32_t MLNKVReadUInt32(uint8_t* ptr, int16_t size) {
+static inline uint32_t MLNKVReadUInt32(uint8_t* ptr, uint8_t size) {
     uint32_t value = 0;
-    for (int16_t i = 0; i < size; i++) {
+    for (uint8_t i = 0; i < size; i++) {
         value |= (uint32_t)(ptr[i] & 0xFF) << (8 * i);
     }
     return value;
 }
 
-static inline uint64_t MLNKVReadUInt64(uint8_t* ptr, int16_t size) {
+static inline uint64_t MLNKVReadUInt64(uint8_t* ptr, uint8_t size) {
     uint64_t value = 0;
-    for (int16_t i = 0; i < size; i++) {
+    for (uint8_t i = 0; i < size; i++) {
         value |= (uint64_t)(ptr[i] & 0xFF) << (8 * i);
     }
     return value;
 }
 
-static inline bool MLNKVWriteUInt32(uint8_t* ptr, uint32_t value, int16_t size) {
-    for (int16_t i = 0; i < size; i++) {
+static inline bool MLNKVWriteUInt32(uint8_t* ptr, uint32_t value, uint8_t size) {
+    for (uint8_t i = 0; i < size; i++) {
         ptr[i] = (uint8_t)(value & 0xFF);
         value = value >> 8;
     }
     return true;
 }
 
-static inline bool MLNKVWriteUInt64(uint8_t* ptr, uint64_t value, int16_t size) {
-    for (int16_t i = 0; i < size; i++) {
+static inline bool MLNKVWriteUInt64(uint8_t* ptr, uint64_t value, uint8_t size) {
+    for (uint8_t i = 0; i < size; i++) {
         ptr[i] = (uint8_t)(value & 0xFF);
         value = value >> 8;
     }
